+++ boilerplate/scripts/build.js
@@ -3,7 +3,6 @@ import builder from "statinamic/lib/builder"
 import config from "./config.js"
 
 const exports = {
-  layouts: require.resolve("../web_modules/layouts"),
   metadata: require.resolve("../web_modules/app/metadata"),
   routes: require.resolve("../web_modules/app/routes"),
 }