+++ boilerplate/web_modules/app/routes.js
@@ -1,11 +1,29 @@
-import React from "react"
+import React, { Component } from "react"
 import { Route } from "react-router"
-import PageContainer from "statinamic/lib/PageContainer"
 
-// Layouts
 import LayoutContainer from "../LayoutContainer"
+import StatinamicPageContainer from "statinamic/lib/PageContainer"
 
-// routes
+import Page from "../layouts/Page"
+import PageError from "../layouts/PageError"
+import PageLoading from "../layouts/PageLoading"
+
+class PageContainer extends Component {
+  render() {
+    const { props } = this
+    return (
+      <StatinamicPageContainer
+        { ...props }
+        layouts={ {
+          Page,
+          PageError,
+          PageLoading,
+        } }
+      />
+    )
+  }
+}
+
 export default (
   <Route component={ LayoutContainer }>
     <Route path="*" component={ PageContainer } />