+++ docs/web_modules/layouts/index.js
@@ -1,6 +1,0 @@
-export Page from "./Page" // default layout
-export PageError from "./PageError"
-
-export Homepage from "./Homepage"
-// you can add your own layout here
-// eg: export Post from "Post"