+++ scripts/docs.js
@@ -1,3 +1,3 @@
 import testFolder from "./utils/test-folder.js"
 
-testFolder("docs")
+testFolder("docs", { lib: "src" })